--- src/main/java/com/sfa/volunteer/dto/CreateUserRequest.java
@@ -1,13 +1,0 @@
-package com.sfa.volunteer.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Builder;
-
-@Builder
-public record CreateUserRequest(
-        @NotBlank String name,
-        @NotBlank @Email String email,
-        @NotBlank String phoneNumber,
-        @NotBlank String timeZone) {
-}
