--- src/main/java/com/sfa/volunteer/dto/PaginationResponse.java
@@ -1,17 +1,0 @@
-package com.sfa.volunteer.dto;
-
-import lombok.Builder;
-
-import java.util.List;
-
-@Builder
-public record PaginationResponse<T>(
-        int currentPage,
-        int pageSize,
-        int totalPages,
-        long totalItems,
-        List<T> items,
-        boolean hasNextPage,
-        boolean hasPreviousPage) {
-}
-
