--- src/main/java/com/sfa/volunteer/model/UserCategory.java
@@ -1,37 +1,0 @@
-package com.sfa.volunteer.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.ZonedDateTime;
-
-@Entity
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "user_category")
-public class UserCategory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "user_category_id")
-    private Integer userCategoryId;
-
-    @Column(name = "user_category")
-    private String userCategory;
-
-    @Column(name = "user_category_desc")
-    private String userCategoryDesc;
-
-    @Column(name = "last_update_date")
-    private ZonedDateTime lastUpdateDate;
-}
