--- src/main/java/com/sfa/volunteer/repository/CountryRepository.java
@@ -1,9 +1,0 @@
-package com.sfa.volunteer.repository;
-
-import com.sfa.volunteer.model.Country;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CountryRepository extends JpaRepository<Country, Integer> {
-}
