--- src/main/java/com/sfa/volunteer/repository/StateRepository.java
@@ -1,9 +1,0 @@
-package com.sfa.volunteer.repository;
-
-import com.sfa.volunteer.model.State;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StateRepository extends JpaRepository<State, Integer> {
-}
