--- src/main/java/com/sfa/volunteer/repository/UserCategoryRepository.java
@@ -1,9 +1,0 @@
-package com.sfa.volunteer.repository;
-
-import com.sfa.volunteer.model.UserCategory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserCategoryRepository extends JpaRepository<UserCategory, Integer> {
-}
