--- src/main/java/com/sfa/volunteer/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.sfa.volunteer.repository;
-
-import com.sfa.volunteer.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, String> {
-}
