--- src/main/java/com/sfa/volunteer/repository/UserStatusRepository.java
@@ -1,9 +1,0 @@
-package com.sfa.volunteer.repository;
-
-import com.sfa.volunteer.model.UserStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserStatusRepository extends JpaRepository<UserStatus, Integer> {
-}
