+++ src/main/java/org/sfa/volunteer/VolunteerApplication.java
@@ -1,0 +1,13 @@
+package org.sfa.volunteer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class VolunteerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(VolunteerApplication.class, args);
+    }
+
+}
