+++ src/main/java/org/sfa/volunteer/dto/CreateUserRequest.java
@@ -1,0 +1,13 @@
+package org.sfa.volunteer.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Builder;
+
+@Builder
+public record CreateUserRequest(
+        @NotBlank String name,
+        @NotBlank @Email String email,
+        @NotBlank String phoneNumber,
+        @NotBlank String timeZone) {
+}
