+++ src/main/java/org/sfa/volunteer/dto/PaginationResponse.java
@@ -1,0 +1,17 @@
+package org.sfa.volunteer.dto;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record PaginationResponse<T>(
+        int currentPage,
+        int pageSize,
+        int totalPages,
+        long totalItems,
+        List<T> items,
+        boolean hasNextPage,
+        boolean hasPreviousPage) {
+}
+
