+++ src/main/java/org/sfa/volunteer/repository/CountryRepository.java
@@ -1,0 +1,9 @@
+package org.sfa.volunteer.repository;
+
+import org.sfa.volunteer.model.Country;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CountryRepository extends JpaRepository<Country, Integer> {
+}
