+++ src/main/java/org/sfa/volunteer/repository/StateRepository.java
@@ -1,0 +1,9 @@
+package org.sfa.volunteer.repository;
+
+import org.sfa.volunteer.model.State;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StateRepository extends JpaRepository<State, Integer> {
+}
