+++ src/main/java/org/sfa/volunteer/repository/UserCategoryRepository.java
@@ -1,0 +1,9 @@
+package org.sfa.volunteer.repository;
+
+import org.sfa.volunteer.model.UserCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserCategoryRepository extends JpaRepository<UserCategory, Integer> {
+}
