+++ src/main/java/org/sfa/volunteer/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package org.sfa.volunteer.repository;
+
+import org.sfa.volunteer.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+}
