+++ src/main/java/org/sfa/volunteer/repository/UserStatusRepository.java
@@ -1,0 +1,9 @@
+package org.sfa.volunteer.repository;
+
+import org.sfa.volunteer.model.UserStatus;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserStatusRepository extends JpaRepository<UserStatus, Integer> {
+}
