--- src/test/java/com/sfa/volunteer/VolunteerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sfa.volunteer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class VolunteerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
