+++ src/test/java/org/sfa/volunteer/VolunteerApplicationTests.java
@@ -1,0 +1,13 @@
+package org.sfa.volunteer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VolunteerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
